User login Attempted Second

+++ src/utils/cloudinaryHandler.js
@@ -15,13 +15,15 @@ const cloudinaryUpload = async (localFilePath) => {
         const response = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto"
         })
-        // file has been uploaded successfull
+        // file has been uploaded successfully
         //console.log("file is uploaded on cloudinary ", response.url);
-        fs.unlinkSync(localFilePath)
+        if (response) {
+            fs.unlinkSync(localFilePath)
+        }
         return response;
 
     } catch (error) {
-        // console.log(error)
+        console.log(error)
         fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
         return null;
     }
